% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_entropy_optimization.R
\name{categorize}
\alias{categorize}
\title{4 categories.}
\usage{
categorize(cutoffs, data)
}
\arguments{
\item{cutoffs}{Discrete bins.}

\item{data}{Numeric vector. The degrees of entropy to calculate.}
}
\value{
Data frame of discretized variables.
}
\description{
4 categories.
}
\examples{
\\code(categorize(c(0, 0, 0.2, 0.4), df))
}
