% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_entropy_optimization.R
\name{entropy_table}
\alias{entropy_table}
\title{A table of entropy measures}
\usage{
entropy_table(entropy_list)
}
\arguments{
\item{entropy_list}{Takes the output of the \code{multi_entropy()} function as input a list of data frames
containing univariate and bivariate entropies.}
}
\value{
Returns a data frame of entropy measures: univariate entropies,
bivariate entropies, mutual information and dependence measures.
}
\description{
A table of entropy measures
}
\examples{
\dontrun{
df_list <- multi_entropy(df, n_variate = 1:2)
df_entropies <- entropy_table(df_list)}
}
