% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_entropy_optimization.R
\name{merge_entropy_table}
\alias{merge_entropy_table}
\title{Title}
\usage{
merge_entropy_table(data)
}
\arguments{
\item{data}{Data frame with discrete variables.}
}
\value{
Merges the list of lists from the \code{multi_entropy()} function.
}
\description{
Title
}
