% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_entropy_optimization.R
\name{multi_entropy}
\alias{multi_entropy}
\title{Calculate n-variate entropy}
\usage{
multi_entropy(data, n_variate = 2)
}
\arguments{
\item{data}{Data frame containing only discrete data.}

\item{n_variate}{Numeric vector. The degrees of entropy to calculate.}
}
\value{
Returns a list of lists with each variables' n-variate entropies.
}
\description{
Calculate n-variate entropy
}
\examples{
\dontrun{
multi_entropy(df, n_variate = 1:3)}
}
