% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_entropy_optimization.R
\name{optimize_entropy}
\alias{optimize_entropy}
\title{Find bin sizes which optimize the entropy}
\usage{
optimize_entropy(data)
}
\arguments{
\item{data}{}
}
\value{
Returns an optim() object
}
\description{
Find bin sizes which optimize the entropy
}
\examples{
optimize_entropy(df)
}
