% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_entropy_optimization.R
\name{sum_entropy}
\alias{sum_entropy}
\title{Calculate the sum of entropies from all variables}
\usage{
sum_entropy(cutoffs, data)
}
\arguments{
\item{cutoffs}{Initial values for cutoffs.}

\item{data}{Data frame containing only discrete data.}
}
\value{
Numeric. Sum of all variables' entropies.
}
\description{
Calculate the sum of entropies from all variables
}
